version 0.5.2
after bloating the version 0.5.1 i reseted from master rep to push the
latest version.

+++ honeymoon/car.pde
@@ -7,14 +7,13 @@ void updateCar() {
   carVx+=carAx;
   carVx*=0.95;
   carVy*=0.95;
-
   carX+=carVx;
   carY+=carVy;
   if (carY< groundL-carH) {
     carY*=1.02;
-  }else{ //onGround
-  carVy=0;
-  onGround=true;
+  } else { //onGround
+    carVy=0;
+    onGround=true;
   }
 }
 
@@ -25,14 +24,14 @@ void displayCar(float x, float y, float w, float h) {  // funktion för bilen
   imageMode(CORNER);
   image(car, int(x+carVibrationOffsetX), int(y+carVibrationOffsetY), w, h);  // bil chassi
 
-  if( onGround && slowFactor!=0)particles.add(new particle(0, x+w-wheelSize, y+h+wheelSize/4, -5+random(5), -3+random(2), random(360),50));  // skapar rök partiklar
+  if ( onGround && slowFactor!=0)particles.add(new particle(0, int(x+w-wheelSize), int(y+h+wheelSize/4), -5+random(5), -3+random(2), random(360), 50));  // skapar rök partiklar
   pushMatrix();
   translate( x+w-wheelOffset*8+random(2) +wheelSize/2, y+h-wheelSize/2+random(2)+wheelSize/2);            // front wheel
   rotate(radians(angle));
   image(wheel1, 0-wheelSize/2, 0-wheelSize/2, wheelSize, wheelSize);  // hjul F
   popMatrix();
 
-  if( onGround && slowFactor!=0)particles.add(new particle(0, x+wheelSize/2, y+h+wheelSize/4, -5+random(5), -3+random(2), random(360),50));  // skapar rök partiklar
+  if ( onGround && slowFactor!=0)particles.add(new particle(0, int(x+wheelSize/2), int(y+h+wheelSize/4), -5+random(5), -3+random(2), random(360), 50));  // skapar rök partiklar
   pushMatrix();
   translate( x+wheelOffset+wheelSize/2+random(2), y+h-wheelSize/2+random(2)+wheelSize/2);            // front wheel
   rotate(radians(angle));
